stm32f4xx: clocks: refactoring: move peripheral clock struct to new module
Peripheral clock logic moved from rcc.rs to a new clocks/periph.rs module.
All peripheral modules adapted to access the peripheral clock though the
new module path.
