+++ chips/stm32f4xx/src/rcc.rs
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 // Copyright Tock Contributors 2022.
 
-use kernel::platform::chip::ClockInterface;
 use kernel::utilities::registers::interfaces::{ReadWriteable, Readable};
 use kernel::utilities::registers::{register_bitfields, ReadWrite};
 use kernel::utilities::StaticRef;
@@ -1054,337 +1053,340 @@ impl Rcc {
         }
     }
 
-    fn configure_rng_clock(&self) {
+    pub(crate) fn configure_rng_clock(&self) {
         self.registers.pllcfgr.modify(PLLCFGR::PLLQ.val(2));
         self.registers.cr.modify(CR::PLLON::SET);
     }
 
     // I2C1 clock
 
-    fn is_enabled_i2c1_clock(&self) -> bool {
+    pub(crate) fn is_enabled_i2c1_clock(&self) -> bool {
         self.registers.apb1enr.is_set(APB1ENR::I2C1EN)
     }
 
-    fn enable_i2c1_clock(&self) {
+    pub(crate) fn enable_i2c1_clock(&self) {
         self.registers.apb1enr.modify(APB1ENR::I2C1EN::SET);
         self.registers.apb1rstr.modify(APB1RSTR::I2C1RST::SET);
         self.registers.apb1rstr.modify(APB1RSTR::I2C1RST::CLEAR);
     }
 
-    fn disable_i2c1_clock(&self) {
+    pub(crate) fn disable_i2c1_clock(&self) {
         self.registers.apb1enr.modify(APB1ENR::I2C1EN::CLEAR)
     }
 
     // SPI3 clock
 
-    fn is_enabled_spi3_clock(&self) -> bool {
+    pub(crate) fn is_enabled_spi3_clock(&self) -> bool {
         self.registers.apb1enr.is_set(APB1ENR::SPI3EN)
     }
 
-    fn enable_spi3_clock(&self) {
+    pub(crate) fn enable_spi3_clock(&self) {
         self.registers.apb1enr.modify(APB1ENR::SPI3EN::SET)
     }
 
-    fn disable_spi3_clock(&self) {
+    pub(crate) fn disable_spi3_clock(&self) {
         self.registers.apb1enr.modify(APB1ENR::SPI3EN::CLEAR)
     }
 
     // TIM2 clock
+    pub(crate) fn is_enabled_tim_pre(&self) -> bool {
+        self.registers.dckcfgr.is_set(DCKCFGR::TIMPRE)
+    }
 
-    fn is_enabled_tim2_clock(&self) -> bool {
+    pub(crate) fn is_enabled_tim2_clock(&self) -> bool {
         self.registers.apb1enr.is_set(APB1ENR::TIM2EN)
     }
 
-    fn enable_tim2_clock(&self) {
+    pub(crate) fn enable_tim2_clock(&self) {
         self.registers.apb1enr.modify(APB1ENR::TIM2EN::SET)
     }
 
-    fn disable_tim2_clock(&self) {
+    pub(crate) fn disable_tim2_clock(&self) {
         self.registers.apb1enr.modify(APB1ENR::TIM2EN::CLEAR)
     }
 
     // SYSCFG clock
 
-    fn is_enabled_syscfg_clock(&self) -> bool {
+    pub(crate) fn is_enabled_syscfg_clock(&self) -> bool {
         self.registers.apb2enr.is_set(APB2ENR::SYSCFGEN)
     }
 
-    fn enable_syscfg_clock(&self) {
+    pub(crate) fn enable_syscfg_clock(&self) {
         self.registers.apb2enr.modify(APB2ENR::SYSCFGEN::SET)
     }
 
-    fn disable_syscfg_clock(&self) {
+    pub(crate) fn disable_syscfg_clock(&self) {
         self.registers.apb2enr.modify(APB2ENR::SYSCFGEN::CLEAR)
     }
 
     // DMA1 clock
 
-    fn is_enabled_dma1_clock(&self) -> bool {
+    pub(crate) fn is_enabled_dma1_clock(&self) -> bool {
         self.registers.ahb1enr.is_set(AHB1ENR::DMA1EN)
     }
 
-    fn enable_dma1_clock(&self) {
+    pub(crate) fn enable_dma1_clock(&self) {
         self.registers.ahb1enr.modify(AHB1ENR::DMA1EN::SET)
     }
 
-    fn disable_dma1_clock(&self) {
+    pub(crate) fn disable_dma1_clock(&self) {
         self.registers.ahb1enr.modify(AHB1ENR::DMA1EN::CLEAR)
     }
 
     // DMA2 clock
-    fn is_enabled_dma2_clock(&self) -> bool {
+    pub(crate) fn is_enabled_dma2_clock(&self) -> bool {
         self.registers.ahb1enr.is_set(AHB1ENR::DMA2EN)
     }
 
-    fn enable_dma2_clock(&self) {
+    pub(crate) fn enable_dma2_clock(&self) {
         self.registers.ahb1enr.modify(AHB1ENR::DMA2EN::SET)
     }
 
-    fn disable_dma2_clock(&self) {
+    pub(crate) fn disable_dma2_clock(&self) {
         self.registers.ahb1enr.modify(AHB1ENR::DMA2EN::CLEAR)
     }
 
     // GPIOH clock
 
-    fn is_enabled_gpioh_clock(&self) -> bool {
+    pub(crate) fn is_enabled_gpioh_clock(&self) -> bool {
         self.registers.ahb1enr.is_set(AHB1ENR::GPIOHEN)
     }
 
-    fn enable_gpioh_clock(&self) {
+    pub(crate) fn enable_gpioh_clock(&self) {
         self.registers.ahb1enr.modify(AHB1ENR::GPIOHEN::SET)
     }
 
-    fn disable_gpioh_clock(&self) {
+    pub(crate) fn disable_gpioh_clock(&self) {
         self.registers.ahb1enr.modify(AHB1ENR::GPIOHEN::CLEAR)
     }
 
     // GPIOG clock
 
-    fn is_enabled_gpiog_clock(&self) -> bool {
+    pub(crate) fn is_enabled_gpiog_clock(&self) -> bool {
         self.registers.ahb1enr.is_set(AHB1ENR::GPIOGEN)
     }
 
-    fn enable_gpiog_clock(&self) {
+    pub(crate) fn enable_gpiog_clock(&self) {
         self.registers.ahb1enr.modify(AHB1ENR::GPIOGEN::SET)
     }
 
-    fn disable_gpiog_clock(&self) {
+    pub(crate) fn disable_gpiog_clock(&self) {
         self.registers.ahb1enr.modify(AHB1ENR::GPIOGEN::CLEAR)
     }
 
     // GPIOF clock
 
-    fn is_enabled_gpiof_clock(&self) -> bool {
+    pub(crate) fn is_enabled_gpiof_clock(&self) -> bool {
         self.registers.ahb1enr.is_set(AHB1ENR::GPIOFEN)
     }
 
-    fn enable_gpiof_clock(&self) {
+    pub(crate) fn enable_gpiof_clock(&self) {
         self.registers.ahb1enr.modify(AHB1ENR::GPIOFEN::SET)
     }
 
-    fn disable_gpiof_clock(&self) {
+    pub(crate) fn disable_gpiof_clock(&self) {
         self.registers.ahb1enr.modify(AHB1ENR::GPIOFEN::CLEAR)
     }
 
     // GPIOE clock
 
-    fn is_enabled_gpioe_clock(&self) -> bool {
+    pub(crate) fn is_enabled_gpioe_clock(&self) -> bool {
         self.registers.ahb1enr.is_set(AHB1ENR::GPIOEEN)
     }
 
-    fn enable_gpioe_clock(&self) {
+    pub(crate) fn enable_gpioe_clock(&self) {
         self.registers.ahb1enr.modify(AHB1ENR::GPIOEEN::SET)
     }
 
-    fn disable_gpioe_clock(&self) {
+    pub(crate) fn disable_gpioe_clock(&self) {
         self.registers.ahb1enr.modify(AHB1ENR::GPIOEEN::CLEAR)
     }
 
     // GPIOD clock
 
-    fn is_enabled_gpiod_clock(&self) -> bool {
+    pub(crate) fn is_enabled_gpiod_clock(&self) -> bool {
         self.registers.ahb1enr.is_set(AHB1ENR::GPIODEN)
     }
 
-    fn enable_gpiod_clock(&self) {
+    pub(crate) fn enable_gpiod_clock(&self) {
         self.registers.ahb1enr.modify(AHB1ENR::GPIODEN::SET)
     }
 
-    fn disable_gpiod_clock(&self) {
+    pub(crate) fn disable_gpiod_clock(&self) {
         self.registers.ahb1enr.modify(AHB1ENR::GPIODEN::CLEAR)
     }
 
     // GPIOC clock
 
-    fn is_enabled_gpioc_clock(&self) -> bool {
+    pub(crate) fn is_enabled_gpioc_clock(&self) -> bool {
         self.registers.ahb1enr.is_set(AHB1ENR::GPIOCEN)
     }
 
-    fn enable_gpioc_clock(&self) {
+    pub(crate) fn enable_gpioc_clock(&self) {
         self.registers.ahb1enr.modify(AHB1ENR::GPIOCEN::SET)
     }
 
-    fn disable_gpioc_clock(&self) {
+    pub(crate) fn disable_gpioc_clock(&self) {
         self.registers.ahb1enr.modify(AHB1ENR::GPIOCEN::CLEAR)
     }
 
     // GPIOB clock
 
-    fn is_enabled_gpiob_clock(&self) -> bool {
+    pub(crate) fn is_enabled_gpiob_clock(&self) -> bool {
         self.registers.ahb1enr.is_set(AHB1ENR::GPIOBEN)
     }
 
-    fn enable_gpiob_clock(&self) {
+    pub(crate) fn enable_gpiob_clock(&self) {
         self.registers.ahb1enr.modify(AHB1ENR::GPIOBEN::SET)
     }
 
-    fn disable_gpiob_clock(&self) {
+    pub(crate) fn disable_gpiob_clock(&self) {
         self.registers.ahb1enr.modify(AHB1ENR::GPIOBEN::CLEAR)
     }
 
     // GPIOA clock
 
-    fn is_enabled_gpioa_clock(&self) -> bool {
+    pub(crate) fn is_enabled_gpioa_clock(&self) -> bool {
         self.registers.ahb1enr.is_set(AHB1ENR::GPIOAEN)
     }
 
-    fn enable_gpioa_clock(&self) {
+    pub(crate) fn enable_gpioa_clock(&self) {
         self.registers.ahb1enr.modify(AHB1ENR::GPIOAEN::SET)
     }
 
-    fn disable_gpioa_clock(&self) {
+    pub(crate) fn disable_gpioa_clock(&self) {
         self.registers.ahb1enr.modify(AHB1ENR::GPIOAEN::CLEAR)
     }
 
     // FMC
 
-    fn is_enabled_fmc_clock(&self) -> bool {
+    pub(crate) fn is_enabled_fmc_clock(&self) -> bool {
         self.registers.ahb3enr.is_set(AHB3ENR::FMCEN)
     }
 
-    fn enable_fmc_clock(&self) {
+    pub(crate) fn enable_fmc_clock(&self) {
         self.registers.ahb3enr.modify(AHB3ENR::FMCEN::SET)
     }
 
-    fn disable_fmc_clock(&self) {
+    pub(crate) fn disable_fmc_clock(&self) {
         self.registers.ahb3enr.modify(AHB3ENR::FMCEN::CLEAR)
     }
 
     // USART1 clock
-    fn is_enabled_usart1_clock(&self) -> bool {
+    pub(crate) fn is_enabled_usart1_clock(&self) -> bool {
         self.registers.apb2enr.is_set(APB2ENR::USART1EN)
     }
 
-    fn enable_usart1_clock(&self) {
+    pub(crate) fn enable_usart1_clock(&self) {
         self.registers.apb2enr.modify(APB2ENR::USART1EN::SET)
     }
 
-    fn disable_usart1_clock(&self) {
+    pub(crate) fn disable_usart1_clock(&self) {
         self.registers.apb2enr.modify(APB2ENR::USART1EN::CLEAR)
     }
 
     // USART2 clock
 
-    fn is_enabled_usart2_clock(&self) -> bool {
+    pub(crate) fn is_enabled_usart2_clock(&self) -> bool {
         self.registers.apb1enr.is_set(APB1ENR::USART2EN)
     }
 
-    fn enable_usart2_clock(&self) {
+    pub(crate) fn enable_usart2_clock(&self) {
         self.registers.apb1enr.modify(APB1ENR::USART2EN::SET)
     }
 
-    fn disable_usart2_clock(&self) {
+    pub(crate) fn disable_usart2_clock(&self) {
         self.registers.apb1enr.modify(APB1ENR::USART2EN::CLEAR)
     }
 
     // USART3 clock
 
-    fn is_enabled_usart3_clock(&self) -> bool {
+    pub(crate) fn is_enabled_usart3_clock(&self) -> bool {
         self.registers.apb1enr.is_set(APB1ENR::USART3EN)
     }
 
-    fn enable_usart3_clock(&self) {
+    pub(crate) fn enable_usart3_clock(&self) {
         self.registers.apb1enr.modify(APB1ENR::USART3EN::SET)
     }
 
-    fn disable_usart3_clock(&self) {
+    pub(crate) fn disable_usart3_clock(&self) {
         self.registers.apb1enr.modify(APB1ENR::USART3EN::CLEAR)
     }
 
     // ADC1 clock
 
-    fn is_enabled_adc1_clock(&self) -> bool {
+    pub(crate) fn is_enabled_adc1_clock(&self) -> bool {
         self.registers.apb2enr.is_set(APB2ENR::ADC1EN)
     }
 
-    fn enable_adc1_clock(&self) {
+    pub(crate) fn enable_adc1_clock(&self) {
         self.registers.apb2enr.modify(APB2ENR::ADC1EN::SET)
     }
 
-    fn disable_adc1_clock(&self) {
+    pub(crate) fn disable_adc1_clock(&self) {
         self.registers.apb2enr.modify(APB2ENR::ADC1EN::CLEAR)
     }
 
     // DAC clock
 
-    fn is_enabled_dac_clock(&self) -> bool {
+    pub(crate) fn is_enabled_dac_clock(&self) -> bool {
         self.registers.apb1enr.is_set(APB1ENR::DACEN)
     }
 
-    fn enable_dac_clock(&self) {
+    pub(crate) fn enable_dac_clock(&self) {
         self.registers.apb1enr.modify(APB1ENR::DACEN::SET)
     }
 
-    fn disable_dac_clock(&self) {
+    pub(crate) fn disable_dac_clock(&self) {
         self.registers.apb1enr.modify(APB1ENR::DACEN::CLEAR)
     }
 
     // RNG clock
 
-    fn is_enabled_rng_clock(&self) -> bool {
+    pub(crate) fn is_enabled_rng_clock(&self) -> bool {
         self.registers.ahb2enr.is_set(AHB2ENR::RNGEN)
     }
 
-    fn enable_rng_clock(&self) {
+    pub(crate) fn enable_rng_clock(&self) {
         self.registers.ahb2enr.modify(AHB2ENR::RNGEN::SET);
     }
 
-    fn disable_rng_clock(&self) {
+    pub(crate) fn disable_rng_clock(&self) {
         self.registers.ahb2enr.modify(AHB2ENR::RNGEN::CLEAR);
     }
 
     // OTGFS clock
 
-    fn is_enabled_otgfs_clock(&self) -> bool {
+    pub(crate) fn is_enabled_otgfs_clock(&self) -> bool {
         self.registers.ahb2enr.is_set(AHB2ENR::OTGFSEN)
     }
 
-    fn enable_otgfs_clock(&self) {
+    pub(crate) fn enable_otgfs_clock(&self) {
         self.registers.ahb2enr.modify(AHB2ENR::OTGFSEN::SET);
     }
 
-    fn disable_otgfs_clock(&self) {
+    pub(crate) fn disable_otgfs_clock(&self) {
         self.registers.ahb2enr.modify(AHB2ENR::OTGFSEN::CLEAR);
     }
 
     // CAN1 clock
 
-    fn is_enabled_can1_clock(&self) -> bool {
+    pub(crate) fn is_enabled_can1_clock(&self) -> bool {
         self.registers.apb1enr.is_set(APB1ENR::CAN1EN)
     }
 
-    fn enable_can1_clock(&self) {
+    pub(crate) fn enable_can1_clock(&self) {
         self.registers.apb1rstr.modify(APB1RSTR::CAN1RST::SET);
         self.registers.apb1rstr.modify(APB1RSTR::CAN1RST::CLEAR);
         self.registers.apb1enr.modify(APB1ENR::CAN1EN::SET);
     }
 
-    fn disable_can1_clock(&self) {
+    pub(crate) fn disable_can1_clock(&self) {
         self.registers.apb1enr.modify(APB1ENR::CAN1EN::CLEAR);
     }
 
     // RTC clock
-    fn source_into_u32(source: RtcClockSource) -> u32 {
+    pub(crate) fn source_into_u32(source: RtcClockSource) -> u32 {
         match source {
             RtcClockSource::LSE => 1,
             RtcClockSource::LSI => 2,
@@ -1392,28 +1394,28 @@ impl Rcc {
         }
     }
 
-    fn enable_lsi_clock(&self) {
+    pub(crate) fn enable_lsi_clock(&self) {
         self.registers.csr.modify(CSR::LSION::SET);
     }
 
-    fn is_enabled_pwr_clock(&self) -> bool {
+    pub(crate) fn is_enabled_pwr_clock(&self) -> bool {
         self.registers.apb1enr.is_set(APB1ENR::PWREN)
     }
 
-    fn enable_pwr_clock(&self) {
+    pub(crate) fn enable_pwr_clock(&self) {
         // Enable the power interface clock
         self.registers.apb1enr.modify(APB1ENR::PWREN::SET);
     }
 
-    fn disable_pwr_clock(&self) {
+    pub(crate) fn disable_pwr_clock(&self) {
         self.registers.apb1enr.modify(APB1ENR::PWREN::CLEAR);
     }
 
-    fn is_enabled_rtc_clock(&self) -> bool {
+    pub(crate) fn is_enabled_rtc_clock(&self) -> bool {
         self.registers.bdcr.is_set(BDCR::RTCEN)
     }
 
-    fn enable_rtc_clock(&self, source: RtcClockSource) {
+    pub(crate) fn enable_rtc_clock(&self, source: RtcClockSource) {
         // Enable LSI
         self.enable_lsi_clock();
         let mut counter = 1_000;
@@ -1432,7 +1434,7 @@ impl Rcc {
         self.registers.bdcr.modify(BDCR::RTCEN::SET);
     }
 
-    fn disable_rtc_clock(&self) {
+    pub(crate) fn disable_rtc_clock(&self) {
         self.registers.bdcr.modify(BDCR::RTCEN.val(1));
         self.registers.bdcr.modify(BDCR::RTCSEL.val(0));
     }
@@ -1556,335 +1558,6 @@ impl From<APBPrescaler> for usize {
             APBPrescaler::DivideBy4 => 4,
             APBPrescaler::DivideBy8 => 8,
             APBPrescaler::DivideBy16 => 16,
-        }
-    }
-}
-
-use crate::clocks::clocks::Stm32f4Clocks;
-
-pub struct PeripheralClock<'a> {
-    pub clock: PeripheralClockType,
-    clocks: &'a dyn Stm32f4Clocks,
-}
-
-/// Bus + Clock name for the peripherals
-pub enum PeripheralClockType {
-    AHB1(HCLK1),
-    AHB2(HCLK2),
-    AHB3(HCLK3),
-    APB1(PCLK1),
-    APB2(PCLK2),
-    RTC,
-    PWR,
-}
-
-/// Peripherals clocked by HCLK1
-pub enum HCLK1 {
-    DMA1,
-    DMA2,
-    GPIOH,
-    GPIOG,
-    GPIOF,
-    GPIOE,
-    GPIOD,
-    GPIOC,
-    GPIOB,
-    GPIOA,
-}
-
-/// Peripherals clocked by HCLK3
-pub enum HCLK3 {
-    FMC,
-}
-
-/// Peripherals clocked by HCLK2
-pub enum HCLK2 {
-    RNG,
-    OTGFS,
-}
-
-/// Peripherals clocked by PCLK1
-pub enum PCLK1 {
-    TIM2,
-    USART2,
-    USART3,
-    SPI3,
-    I2C1,
-    CAN1,
-    DAC,
-}
-
-/// Peripherals clocked by PCLK2
-pub enum PCLK2 {
-    USART1,
-    ADC1,
-    SYSCFG,
-}
-
-impl<'a> PeripheralClock<'a> {
-    pub const fn new(clock: PeripheralClockType, clocks: &'a dyn Stm32f4Clocks) -> Self {
-        Self { clock, clocks }
-    }
-
-    pub fn configure_rng_clock(&self) {
-        self.clocks.get_rcc().configure_rng_clock();
-    }
-
-    pub fn get_frequency(&self) -> u32 {
-        #[inline(always)]
-        fn tim_freq(rcc: &Rcc, hclk_freq: usize, prescaler: APBPrescaler) -> usize {
-            // Reference Manual RM0090 section 6.2
-            // When TIMPRE bit of the RCC_DCKCFGR register is reset, if APBx prescaler is 1, then
-            // TIMxCLK = PCLKx, otherwise TIMxCLK = 2x PCLKx.
-            // When TIMPRE bit in the RCC_DCKCFGR register is set, if APBx prescaler is 1,2 or 4,
-            // then TIMxCLK = HCLK, otherwise TIMxCLK = 4x PCLKx.
-            if !rcc.registers.dckcfgr.is_set(DCKCFGR::TIMPRE) {
-                match prescaler {
-                    APBPrescaler::DivideBy1 | APBPrescaler::DivideBy2 => hclk_freq,
-                    _ => hclk_freq / usize::from(prescaler) * 2,
-                }
-            } else {
-                match prescaler {
-                    APBPrescaler::DivideBy1 | APBPrescaler::DivideBy2 | APBPrescaler::DivideBy4 => {
-                        hclk_freq
-                    }
-                    _ => hclk_freq / usize::from(prescaler) * 4,
-                }
-            }
-        }
-        let rcc = self.clocks.get_rcc();
-        let hclk_freq = rcc.get_sys_clock_frequency();
-        match self.clock {
-            PeripheralClockType::AHB1(_)
-            | PeripheralClockType::AHB2(_)
-            | PeripheralClockType::AHB3(_) => hclk_freq as u32,
-            PeripheralClockType::APB1(ref v) => {
-                let prescaler = rcc.get_apb1_prescaler();
-                match v {
-                    PCLK1::TIM2 => tim_freq(rcc, hclk_freq, prescaler) as u32,
-                    _ => (hclk_freq / usize::from(prescaler)) as u32,
-                }
-            }
-            PeripheralClockType::APB2(_) => {
-                let prescaler = rcc.get_apb2_prescaler();
-                (hclk_freq / usize::from(prescaler)) as u32
-            }
-            //TODO: implement clock frequency retrieval for RTC and PWR peripherals
-            PeripheralClockType::RTC => todo!(),
-            PeripheralClockType::PWR => todo!(),
-        }
-    }
-}
-
-impl<'a> ClockInterface for PeripheralClock<'a> {
-    fn is_enabled(&self) -> bool {
-        let rcc = self.clocks.get_rcc();
-        match self.clock {
-            PeripheralClockType::AHB1(ref v) => match v {
-                HCLK1::DMA1 => rcc.is_enabled_dma1_clock(),
-                HCLK1::DMA2 => rcc.is_enabled_dma2_clock(),
-                HCLK1::GPIOH => rcc.is_enabled_gpioh_clock(),
-                HCLK1::GPIOG => rcc.is_enabled_gpiog_clock(),
-                HCLK1::GPIOF => rcc.is_enabled_gpiof_clock(),
-                HCLK1::GPIOE => rcc.is_enabled_gpioe_clock(),
-                HCLK1::GPIOD => rcc.is_enabled_gpiod_clock(),
-                HCLK1::GPIOC => rcc.is_enabled_gpioc_clock(),
-                HCLK1::GPIOB => rcc.is_enabled_gpiob_clock(),
-                HCLK1::GPIOA => rcc.is_enabled_gpioa_clock(),
-            },
-            PeripheralClockType::AHB2(ref v) => match v {
-                HCLK2::RNG => rcc.is_enabled_rng_clock(),
-                HCLK2::OTGFS => rcc.is_enabled_otgfs_clock(),
-            },
-            PeripheralClockType::AHB3(ref v) => match v {
-                HCLK3::FMC => rcc.is_enabled_fmc_clock(),
-            },
-            PeripheralClockType::APB1(ref v) => match v {
-                PCLK1::TIM2 => rcc.is_enabled_tim2_clock(),
-                PCLK1::USART2 => rcc.is_enabled_usart2_clock(),
-                PCLK1::USART3 => rcc.is_enabled_usart3_clock(),
-                PCLK1::I2C1 => rcc.is_enabled_i2c1_clock(),
-                PCLK1::SPI3 => rcc.is_enabled_spi3_clock(),
-                PCLK1::CAN1 => rcc.is_enabled_can1_clock(),
-                PCLK1::DAC => rcc.is_enabled_dac_clock(),
-            },
-            PeripheralClockType::APB2(ref v) => match v {
-                PCLK2::USART1 => rcc.is_enabled_usart1_clock(),
-                PCLK2::ADC1 => rcc.is_enabled_adc1_clock(),
-                PCLK2::SYSCFG => rcc.is_enabled_syscfg_clock(),
-            },
-            PeripheralClockType::RTC => rcc.is_enabled_rtc_clock(),
-            PeripheralClockType::PWR => rcc.is_enabled_pwr_clock(),
-        }
-    }
-
-    fn enable(&self) {
-        let rcc = self.clocks.get_rcc();
-        match self.clock {
-            PeripheralClockType::AHB1(ref v) => match v {
-                HCLK1::DMA1 => {
-                    rcc.enable_dma1_clock();
-                }
-                HCLK1::DMA2 => {
-                    rcc.enable_dma2_clock();
-                }
-                HCLK1::GPIOH => {
-                    rcc.enable_gpioh_clock();
-                }
-                HCLK1::GPIOG => {
-                    rcc.enable_gpiog_clock();
-                }
-                HCLK1::GPIOF => {
-                    rcc.enable_gpiof_clock();
-                }
-                HCLK1::GPIOE => {
-                    rcc.enable_gpioe_clock();
-                }
-                HCLK1::GPIOD => {
-                    rcc.enable_gpiod_clock();
-                }
-                HCLK1::GPIOC => {
-                    rcc.enable_gpioc_clock();
-                }
-                HCLK1::GPIOB => {
-                    rcc.enable_gpiob_clock();
-                }
-                HCLK1::GPIOA => {
-                    rcc.enable_gpioa_clock();
-                }
-            },
-            PeripheralClockType::AHB2(ref v) => match v {
-                HCLK2::RNG => {
-                    rcc.enable_rng_clock();
-                }
-                HCLK2::OTGFS => {
-                    rcc.enable_otgfs_clock();
-                }
-            },
-            PeripheralClockType::AHB3(ref v) => match v {
-                HCLK3::FMC => rcc.enable_fmc_clock(),
-            },
-            PeripheralClockType::APB1(ref v) => match v {
-                PCLK1::TIM2 => {
-                    rcc.enable_tim2_clock();
-                }
-                PCLK1::USART2 => {
-                    rcc.enable_usart2_clock();
-                }
-                PCLK1::USART3 => {
-                    rcc.enable_usart3_clock();
-                }
-                PCLK1::I2C1 => {
-                    rcc.enable_i2c1_clock();
-                }
-                PCLK1::SPI3 => {
-                    rcc.enable_spi3_clock();
-                }
-                PCLK1::CAN1 => {
-                    rcc.enable_can1_clock();
-                }
-                PCLK1::DAC => {
-                    rcc.enable_dac_clock();
-                }
-            },
-            PeripheralClockType::APB2(ref v) => match v {
-                PCLK2::USART1 => {
-                    rcc.enable_usart1_clock();
-                }
-                PCLK2::ADC1 => {
-                    rcc.enable_adc1_clock();
-                }
-                PCLK2::SYSCFG => {
-                    rcc.enable_syscfg_clock();
-                }
-            },
-            PeripheralClockType::RTC => rcc.enable_rtc_clock(RtcClockSource::LSI),
-            PeripheralClockType::PWR => rcc.enable_pwr_clock(),
-        }
-    }
-
-    fn disable(&self) {
-        let rcc = self.clocks.get_rcc();
-        match self.clock {
-            PeripheralClockType::AHB1(ref v) => match v {
-                HCLK1::DMA1 => {
-                    rcc.disable_dma1_clock();
-                }
-                HCLK1::DMA2 => {
-                    rcc.disable_dma2_clock();
-                }
-                HCLK1::GPIOH => {
-                    rcc.disable_gpioh_clock();
-                }
-                HCLK1::GPIOG => {
-                    rcc.disable_gpiog_clock();
-                }
-                HCLK1::GPIOF => {
-                    rcc.disable_gpiof_clock();
-                }
-                HCLK1::GPIOE => {
-                    rcc.disable_gpioe_clock();
-                }
-                HCLK1::GPIOD => {
-                    rcc.disable_gpiod_clock();
-                }
-                HCLK1::GPIOC => {
-                    rcc.disable_gpioc_clock();
-                }
-                HCLK1::GPIOB => {
-                    rcc.disable_gpiob_clock();
-                }
-                HCLK1::GPIOA => {
-                    rcc.disable_gpioa_clock();
-                }
-            },
-            PeripheralClockType::AHB2(ref v) => match v {
-                HCLK2::RNG => {
-                    rcc.disable_rng_clock();
-                }
-                HCLK2::OTGFS => {
-                    rcc.disable_otgfs_clock();
-                }
-            },
-            PeripheralClockType::AHB3(ref v) => match v {
-                HCLK3::FMC => rcc.disable_fmc_clock(),
-            },
-            PeripheralClockType::APB1(ref v) => match v {
-                PCLK1::TIM2 => {
-                    rcc.disable_tim2_clock();
-                }
-                PCLK1::USART2 => {
-                    rcc.disable_usart2_clock();
-                }
-                PCLK1::USART3 => {
-                    rcc.disable_usart3_clock();
-                }
-                PCLK1::I2C1 => {
-                    rcc.disable_i2c1_clock();
-                }
-                PCLK1::SPI3 => {
-                    rcc.disable_spi3_clock();
-                }
-                PCLK1::CAN1 => {
-                    rcc.disable_can1_clock();
-                }
-                PCLK1::DAC => {
-                    rcc.disable_dac_clock();
-                }
-            },
-            PeripheralClockType::APB2(ref v) => match v {
-                PCLK2::USART1 => {
-                    rcc.disable_usart1_clock();
-                }
-                PCLK2::ADC1 => {
-                    rcc.disable_adc1_clock();
-                }
-                PCLK2::SYSCFG => {
-                    rcc.disable_syscfg_clock();
-                }
-            },
-            PeripheralClockType::RTC => rcc.disable_rtc_clock(),
-            PeripheralClockType::PWR => rcc.disable_pwr_clock(),
         }
     }
 }
