+++ chips/stm32f4xx/src/spi.rs
@@ -15,9 +15,9 @@ use kernel::utilities::registers::interfaces::{ReadWriteable, Readable};
 use kernel::utilities::registers::{register_bitfields, ReadOnly, ReadWrite};
 use kernel::utilities::StaticRef;
 
+use crate::clocks::phclk;
 use crate::dma;
 use crate::dma::{Dma1, Dma1Peripheral};
-use crate::rcc;
 
 /// Serial peripheral interface
 #[repr(C)]
@@ -510,7 +510,7 @@ impl<'a> dma::StreamClient<'a, Dma1<'a>> for Spi<'a> {
     }
 }
 
-pub struct SpiClock<'a>(pub rcc::PeripheralClock<'a>);
+pub struct SpiClock<'a>(pub phclk::PeripheralClock<'a>);
 
 impl ClockInterface for SpiClock<'_> {
     fn is_enabled(&self) -> bool {
