+++ chips/stm32f4xx/src/trng.rs
@@ -4,8 +4,7 @@
 
 //! True random number generator
 
-use crate::clocks::Stm32f4Clocks;
-use crate::rcc;
+use crate::clocks::{phclk, Stm32f4Clocks};
 use kernel::hil;
 use kernel::hil::entropy::Continue;
 use kernel::platform::chip::ClockInterface;
@@ -65,8 +64,8 @@ impl<'a> Trng<'a> {
     ) -> Trng<'a> {
         Trng {
             registers: registers,
-            clock: RngClock(rcc::PeripheralClock::new(
-                rcc::PeripheralClockType::AHB2(rcc::HCLK2::RNG),
+            clock: RngClock(phclk::PeripheralClock::new(
+                phclk::PeripheralClockType::AHB2(phclk::HCLK2::RNG),
                 clocks,
             )),
             client: OptionalCell::empty(),
@@ -112,7 +111,7 @@ impl<'a> Trng<'a> {
     }
 }
 
-struct RngClock<'a>(rcc::PeripheralClock<'a>);
+struct RngClock<'a>(phclk::PeripheralClock<'a>);
 
 impl ClockInterface for RngClock<'_> {
     fn is_enabled(&self) -> bool {
